HBASE-17950 Write the chunkId also as Int instead of long into the first byte of the chunk (Ram)

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/OffheapChunk.java
@@ -41,7 +41,7 @@ public class OffheapChunk extends Chunk {
   void allocateDataBuffer() {
     if (data == null) {
       data = ByteBuffer.allocateDirect(this.size);
-      data.putLong(0, this.getId());
+      data.putInt(0, this.getId());
     }
   }
 }
